fix(server): prevents a user who is not the owner from deleting the workspace

+++ server/src/domain/usecases/DeleteWorkspace.ts
@@ -1,13 +1,23 @@
 import IDeleteWorkspaceRepository from 'infra/protocols/repositories/IDeleteWorkspaceRepository';
+import ILoadWorkspaceById from 'infra/protocols/repositories/ILoadWorkspaceById';
 
 export default class DeleteWorkspace {
+  private loadWorkspaceById: ILoadWorkspaceById;
   private deleteWorkspaceRepository: IDeleteWorkspaceRepository;
 
-  constructor(deleteWorkspaceRepository: IDeleteWorkspaceRepository) {
+  constructor(
+    loadWorkspaceById: ILoadWorkspaceById,
+    deleteWorkspaceRepository: IDeleteWorkspaceRepository,
+  ) {
+    this.loadWorkspaceById = loadWorkspaceById;
     this.deleteWorkspaceRepository = deleteWorkspaceRepository;
   }
 
   async delete(userId: string, workspaceId: string): Promise<boolean> {
+    const workspace = await this.loadWorkspaceById.loadById(workspaceId);
+    if (!workspace) return false;
+    if (workspace.ownerId !== userId) return false;
+
     return this.deleteWorkspaceRepository.delete(userId, workspaceId);
   }
 }
